+++ amira-denoising/process-results/process-results.sh
@@ -146,6 +146,23 @@ generate_supersets() {
 
 # Processes each image to generate the final collage
 process_results() {
+    # Check orig_dir
+    if [ -z "$orig_dir" ]; do
+        echo "Please specify a non-empty path to retrieve original results"
+        exit
+    done
+
+    # Check proc_dir before removing any content
+    if [ -z "$proc_dir" ]; do
+        echo "Please specify a non-empty path to store processed results"
+        exit
+    done
+
+    if [ "$proc_dir" == "$HOME" ]; do
+        echo "$HOME is not valid destination directory"
+        exit
+    done
+
     # Remove any previous result from the processed directory and copy
     # original images
     rm -rf $proc_dir/*
